Initial commit
Not the first time I'm working on this project. So far the commands to add anime, and anime to the lists of watching for yourself and watching with others are done. You can also update anime in the anime list, not yet in the other lists. I have yet to implement the queues, some updates and all delete methods. Also need to figure out how I can make a build once I'm done.